@@ -25,9 +25,10 @@
   "dependencies": {
     "axios": "^0.12.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "material-ui": "^0.16.4",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "material-ui": "^0.19.0",
+    "material-ui-pagination": "^1.1.6",
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2",
     "react-redux": "^4.4.8",
     "react-router": "^3.0.2",
     "react-tap-event-plugin": "^2.0.1",
