@@ -13,13 +13,9 @@ import {render} from 'react-dom';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 import Main from './app/Main'; // Our custom react component
 
-import { Router, Route, IndexRoute, hashHistory } from "react-router";
-import SecondPage from "./app/SecondPage";
-import FirstPage from "./app/FirstPage";
-import CountryForm from "./app/CountryForm";
-import UserForm from "./app/UserForm";
-import ListTest from "./app/ListTest";
+
 import { Provider } from "react-redux"
+import { Routes } from "/routes";
 import store from "./app/base/store"
 import './www/main.css';
 
@@ -30,18 +26,10 @@ injectTapEventPlugin();
 
 // Render the main app react component into the app div.
 // For more details see: https://facebook.github.io/react/docs/top-level-api.html#react.render
+// Might obsoleted
 render(
     <Provider store={store}>
-        <Router history={hashHistory}>
-            <Route path="/" component={Main}>
-                <IndexRoute component={SecondPage}></IndexRoute>
-                <Route path="secondPage" name="secondPage" component={SecondPage}></Route>
-                <Route path="firstPage" name="firstPage" component={FirstPage}></Route>
-                <Route path="countryForm" name="countryForm" component={CountryForm}></Route>
-                <Route path="userForm" name="userForm" component={UserForm}></Route>
-                <Route path="listTest" name="listTest" component={ListTest}></Route>
-            </Route>
-        </Router>
+        <Routes />
     </Provider>,
     document.getElementById('root')
 );
