@@ -11,6 +11,7 @@ import UserForm from "./UserForm";
 import { Provider } from "react-redux"
 import store from "./base/store"
 import Layout from "./Layout";
+import routes from "./CustomRoutes";
 // Needed for onTouchTap
 // http://stackoverflow.com/a/34015469/988941
 injectTapEventPlugin();
@@ -19,15 +20,16 @@ injectTapEventPlugin();
 // For more details see: https://facebook.github.io/react/docs/top-level-api.html#react.render
 render(
     <Provider store={store}>
-        <Router history={hashHistory}>
-            <Route path="/" component={Main}>
-                <IndexRoute component={SecondPage}></IndexRoute>
-                <Route path="secondPage" name="secondPage" component={SecondPage}></Route>
-                <Route path="firstPage" name="firstPage" component={FirstPage}></Route>
-                <Route path="countryForm" name="countryForm" component={CountryForm}></Route>
-                <Route path="userForm" name="userForm" component={UserForm}></Route>
-            </Route>
-        </Router>
+        <Router history={hashHistory} routes={routes} />
+            {/*<Route path="/" component={Main}>*/}
+                {/*<IndexRoute component={SecondPage}></IndexRoute>*/}
+                {/*<Route path="secondPage" name="secondPage" component={SecondPage}></Route>*/}
+                {/*<Route path="firstPage" name="firstPage" component={FirstPage}></Route>*/}
+                {/*<Route path="countryForm" name="countryForm" component={CountryForm}></Route>*/}
+                {/*<Route path="userForm" name="userForm" component={UserForm}></Route>*/}
+            {/*</Route>*/}
+        {/*</Router>*/}
+
     </Provider>,
     document.getElementById('app')
 );
